refactor: create factories folder
