+++ src/api/research-study/gateways/EsResearchStudyRepository.test.ts
@@ -1,8 +1,8 @@
 import { EsResearchStudyRepository } from './EsResearchStudyRepository'
 import { riphCtisDto, setupClientAndElasticsearchService } from '../../../shared/test/helpers/elasticsearchHelper'
 import { ElasticsearchBodyType } from '../application/entities/ElasticsearchBody'
+import { RiphCtisResearchStudyModelFactory } from 'src/etl/factories/RiphCtisResearchStudyModelFactory'
 import { researchStudyIndexMapping } from 'src/etl/researchStudyIndexMapping'
-import { RiphCtisResearchStudyModelFactory } from 'src/etl/RiphCtisResearchStudyModelFactory'
 
 describe('elasticsearch research study repository', () => {
   describe('retrieve one research study', () => {
