+++ src/etl/factories/RiphCtisResearchStudyModelFactory.ts
@@ -1,18 +1,18 @@
 import { Extension, Identifier, Meta } from 'fhir/r4'
 
-import { RiphCtisDto } from './dto/RiphCtisDto'
-import { ModelUtils } from '../shared/models/custom/ModelUtils'
-import { ReferenceContentsModel } from '../shared/models/custom/ReferenceContentsModel'
-import { CodeableConceptModel } from '../shared/models/fhir/DataTypes/CodeableConceptModel'
-import { IdentifierModel } from '../shared/models/fhir/DataTypes/IdentifierModel'
-import { GroupModel } from '../shared/models/fhir/GroupModel'
-import { LocationModel } from '../shared/models/fhir/LocationModel'
-import { ContactDetailModel } from '../shared/models/fhir/MetadataTypes/ContactDetailModel'
-import { OrganizationModel } from '../shared/models/fhir/OrganizationModel'
-import { RiphStatus, ResearchStudyModel } from '../shared/models/fhir/ResearchStudyModel'
-import { ExtensionModel } from '../shared/models/fhir/SpecialPurposeDataTypes/ExtensionModel'
-import { MetaModel } from '../shared/models/fhir/SpecialPurposeDataTypes/MetaModel'
-import { ReferenceModel } from '../shared/models/fhir/SpecialPurposeDataTypes/ReferenceModel'
+import { ModelUtils } from '../../shared/models/custom/ModelUtils'
+import { ReferenceContentsModel } from '../../shared/models/custom/ReferenceContentsModel'
+import { CodeableConceptModel } from '../../shared/models/fhir/DataTypes/CodeableConceptModel'
+import { IdentifierModel } from '../../shared/models/fhir/DataTypes/IdentifierModel'
+import { GroupModel } from '../../shared/models/fhir/GroupModel'
+import { LocationModel } from '../../shared/models/fhir/LocationModel'
+import { ContactDetailModel } from '../../shared/models/fhir/MetadataTypes/ContactDetailModel'
+import { OrganizationModel } from '../../shared/models/fhir/OrganizationModel'
+import { RiphStatus, ResearchStudyModel } from '../../shared/models/fhir/ResearchStudyModel'
+import { ExtensionModel } from '../../shared/models/fhir/SpecialPurposeDataTypes/ExtensionModel'
+import { MetaModel } from '../../shared/models/fhir/SpecialPurposeDataTypes/MetaModel'
+import { ReferenceModel } from '../../shared/models/fhir/SpecialPurposeDataTypes/ReferenceModel'
+import { RiphCtisDto } from '../dto/RiphCtisDto'
 
 export class RiphCtisResearchStudyModelFactory {
   static create(riphCtisDto: RiphCtisDto): ResearchStudyModel {
